+++ modules/registry-service/modules/modules.v1/api_download.tf
@@ -9,40 +9,20 @@ resource "aws_api_gateway_method" "download_GET" {
 
 
 resource "aws_api_gateway_integration" "download_GET" {
-  rest_api_id = aws_api_gateway_method.download_GET.rest_api_id
-  resource_id = aws_api_gateway_method.download_GET.resource_id
-  http_method = aws_api_gateway_method.download_GET.http_method
-
-  type                    = "AWS"
-  uri                     = "arn:aws:apigateway:${local.region_name}:dynamodb:action/GetItem"
+  rest_api_id             = aws_api_gateway_method.download_GET.rest_api_id
+  resource_id             = aws_api_gateway_method.download_GET.resource_id
+  http_method             = aws_api_gateway_method.download_GET.http_method
   integration_http_method = "POST"
-  credentials             = var.credentials_role_arn
-
-  request_templates = {
-    "application/json" = templatefile("${path.module}/files/download_request.tpl", { dynamo_table_name = var.dynamodb_table_name })
-  }
+  type                    = "AWS_PROXY"
+  uri                     = var.lambda_download_invoke_arn
 }
 
-resource "aws_api_gateway_method_response" "download_GET_200" {
-  rest_api_id = aws_api_gateway_method.download_GET.rest_api_id
-  resource_id = aws_api_gateway_method.download_GET.resource_id
-  http_method = aws_api_gateway_method.download_GET.http_method
-  status_code = "200"
+resource "aws_lambda_permission" "apigw_lambda_download" {
+  statement_id  = "AllowExecutionFromAPIGateway"
+  action        = "lambda:InvokeFunction"
+  function_name = var.lambda_download_name
+  principal     = "apigateway.amazonaws.com"
 
-  response_parameters = {
-    "method.response.header.X-Terraform-Get" = true
-  }
-}
-
-resource "aws_api_gateway_integration_response" "download_GET_200" {
-  rest_api_id = aws_api_gateway_integration.download_GET.rest_api_id
-  resource_id = aws_api_gateway_integration.download_GET.resource_id
-  http_method = aws_api_gateway_integration.download_GET.http_method
-  status_code = aws_api_gateway_method_response.download_GET_200.status_code
-
-  response_parameters = {
-    "method.response.header.X-Terraform-Get" = "integration.response.body.Item.Source.S"
-  }
-
-  response_templates = { "application/json" = file("${path.module}/files/download_response.template") }
+  # More: http://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-control-access-using-iam-policies-to-invoke-api.html
+  source_arn = "arn:aws:execute-api:${local.region_name}:${local.account_id}:${ var.rest_api_id}/*/${aws_api_gateway_method.download_GET.http_method}${aws_api_gateway_resource.download.path}"
 }