+++ modules/registry-service/modules/modules.v1/locals.tf
@@ -9,6 +9,9 @@ locals {
   }
 
   region_name = data.aws_region.region.name
+  account_id = data.aws_caller_identity.current.account_id
 }
 
 data "aws_region" "region" {}
+
+data "aws_caller_identity" "current" {}